Move packages for slightly better UX for consumers
