+++ pkg/kubelet/cri/remote/remote_image.go
