+++ pkg/kubelet/cri/remote/remote_runtime.go
@@ -26,10 +26,10 @@ import (
 	"google.golang.org/grpc"
 	"k8s.io/klog/v2"
 
+	"k8s.io/component-base/logs/logreduction"
 	internalapi "k8s.io/cri-api/pkg/apis"
 	runtimeapi "k8s.io/cri-api/pkg/apis/runtime/v1alpha2"
 	"k8s.io/kubernetes/pkg/kubelet/util"
-	"k8s.io/kubernetes/pkg/kubelet/util/logreduction"
 	utilexec "k8s.io/utils/exec"
 )
