+++ pkg/kubelet/cri/streaming/remotecommand/doc.go
@@ -15,4 +15,4 @@ limitations under the License.
 */
 
 // Package remotecommand contains functions related to executing commands in and attaching to pods.
-package remotecommand // import "k8s.io/kubernetes/pkg/kubelet/server/remotecommand"
+package remotecommand
