+++ pkg/kubelet/cri/streaming/server_test.go
@@ -35,7 +35,7 @@ import (
 	"k8s.io/client-go/tools/remotecommand"
 	"k8s.io/client-go/transport/spdy"
 	runtimeapi "k8s.io/cri-api/pkg/apis/runtime/v1alpha2"
-	kubeletportforward "k8s.io/kubernetes/pkg/kubelet/server/portforward"
+	kubeletportforward "k8s.io/kubernetes/pkg/kubelet/cri/streaming/portforward"
 )
 
 const (
